KSP-87: formatted calendar, added calendar to calendar page
